requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
easyocr==1.7.2
numpy>=1.26.0
opencv-python-headless>=4.8.0
scikit-learn>=1.4.0
Pillow>=10.0.0
pytesseract==0.3.13
requests>=2.31.0
pydantic>=2.5.0
